requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
sendgrid
passlib[bcrypt]
bcrypt
python-multipart
python-jose
email-validator
openai
python-dateutil
alembic
